requirements.txt
Django==1.6.5
MySQL-python==1.2.5
South==1.0
cssutils==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
django-widget-tweaks==1.3
gunicorn==19.0.0
httplib2==0.9
inlinestyler==0.2.0
lxml==3.3.5
parse-rest==0.2.20141004
psycopg2==2.5.3
pyrise==0.5.3
pystache==0.5.4
python-dateutil==2.2
pytz==2014.4
requests==2.3.0
six==1.7.3
static==1.0.2
static3==0.5.1
virtualenv==1.11.6
wsgiref==0.1.2
